// WebClient/src/api/GeneratedApiClient.ts
export interface Conversation {
    maxDuration?: number;
    endingQue?: string | undefined;
    summarization?: Summarization;
    topic?: string | undefined;
    description?: string | undefined;
}

export interface Message {
    text?: string | undefined;
    ssml?: string | undefined;
}

export interface PossibleAnswer {
    type?: string | undefined;
    userPrompt?: string | undefined;
    promptId?: string | undefined;
}

export interface Prompt {
    id?: string | undefined;
    goal?: string | undefined;
    type?: string | undefined;
    message?: Message;
    possibleAnswers?: PossibleAnswer[] | undefined;
    requestedInformation?: RequestedInformation[] | undefined;
    nextPrompt?: string | undefined;
    conversation?: Conversation;
}

export enum PromptTypeEnum {
    User = "User",
    Agent = "Agent",
}

export interface RequestedInformation {
    name?: string | undefined;
    nameEn?: string | undefined;
    description?: string | undefined;
    type?: string | undefined;
    size?: string | undefined;
}

export interface StartRequest {
    newConversation?: boolean;
}

export interface Summarization {
    frequency?: number;
    size?: number;
}

export interface TextToSpeechRequest {
    text?: string | undefined;
}

export interface Unit {
    id?: string | undefined;
    nextUnit?: string | undefined;
    type?: string | undefined;
    name?: string | undefined;
    prompts?: Prompt[] | undefined;
}